+++ include/st_pipeline_api.h
@@ -453,6 +453,12 @@ enum st22_quality_mode {
  * If enabled, simulate random packet loss, test usage only.
  */
 #define ST22P_RX_FLAG_SIMULATE_PKT_LOSS (MTL_BIT32(3))
+/**
+ * Flag bit in flags of struct st22p_rx_ops.
+ * Enable the external frame mode, and user must provide a query callback(query_ext_frame
+ * in st22p_rx_ops) to let MTL can get the frame when needed.
+ */
+#define ST22P_RX_FLAG_EXT_FRAME (MTL_BIT32(4))
 
 /**
  * Flag bit in flags of struct st22p_rx_ops.
@@ -980,6 +986,13 @@ struct st22p_rx_ops {
    * Ex, cast to struct st10_vsync_meta for ST_EVENT_VSYNC.
    */
   int (*notify_event)(void* priv, enum st_event event, void* args);
+  /**
+   * Mandatory for ST22P_RX_FLAG_EXT_FRAME. Callback when the lib query next external
+   * frame's data address. And only non-block method can be used within this callback as
+   * it run from lcore tasklet routine.
+   */
+  int (*query_ext_frame)(void* priv, struct st_ext_frame* ext_frame,
+                         struct st22_rx_frame_meta* meta);
 };
 
 /**